progress

+++ core-web/apps/dotcms-ui/src/app/portlets/dot-experiments/dot-experiments-configuration/dot-experiments-configuration.component.ts
@@ -3,6 +3,8 @@ import { Observable } from 'rxjs';
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 
+import { ConfirmationService } from 'primeng/api';
+
 import { DotMessagePipe } from '@dotcms/app/view/pipes';
 import { DotSessionStorageService } from '@dotcms/data-access';
 import {
@@ -32,7 +34,9 @@ export class DotExperimentsConfigurationComponent implements OnInit {
         private readonly dotExperimentsConfigurationStore: DotExperimentsConfigurationStore,
         private readonly dotSessionStorageService: DotSessionStorageService,
         private readonly router: Router,
-        private readonly route: ActivatedRoute
+        private readonly route: ActivatedRoute,
+        private readonly confirmationService: ConfirmationService,
+        private readonly dotMessagePipe: DotMessagePipe
     ) {}
 
     ngOnInit(): void {
@@ -119,9 +123,17 @@ export class DotExperimentsConfigurationComponent implements OnInit {
      * @memberof DotExperimentsConfigurationComponent
      */
     deleteVariant(variant: Variant, experimentId: string) {
-        this.dotExperimentsConfigurationStore.deleteVariant({
-            experimentId,
-            variant
+        this.confirmationService.confirm({
+            target: event.target,
+            message: this.dotMessagePipe.transform('experiments.configure.variant.delete.confirm'),
+            icon: 'pi pi-exclamation-triangle',
+            acceptLabel: this.dotMessagePipe.transform('delete'),
+            rejectLabel: this.dotMessagePipe.transform('dot.common.dialog.reject'),
+            accept: () =>
+                this.dotExperimentsConfigurationStore.deleteVariant({
+                    experimentId,
+                    variant
+                })
         });
     }
